pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
dependencies = [
    "apexomni",
    "python-dotenv>=1.0.0",
    "requests",
    "web3",
    "ecdsa",
    "mpmath",
    "sympy",
    "python-dateutil",
    "numpy"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"